Split operation files into domain and db
