+++ hedera-mirror-rest/__tests__/specs/transactions-01-no-args.spec.json
@@ -1,85 +1,188 @@
 {
+  "description": "Transaction api calls with no arguments",
+  "setup": {
+    "filename": "transactionsnoarg",
+    "accounts": [
+      {
+        "entity_num": 8
+      },
+      {
+        "entity_num": 9
+      },
+      {
+        "entity_num": 10
+      }
+    ],
+    "balances": [
+      {
+        "timestamp": 3,
+        "id": 8,
+        "balance": 80
+      },
+      {
+        "timestamp": 3,
+        "id": 9,
+        "balance": 90
+      },
+      {
+        "timestamp": 3,
+        "id": 10,
+        "balance": 100
+      }
+    ],
+    "transactions": [],
+    "cryptotransfers": [
+      {
+        "consensus_timestamp": "1234567890000000001",
+        "payerAccountId": "0.0.10",
+        "recipientAccountId": "0.0.9",
+        "amount": 10
+      },
+      {
+        "consensus_timestamp": "1234567890000000002",
+        "payerAccountId": "0.0.10",
+        "recipientAccountId": "0.0.9",
+        "amount": 20
+      },
+      {
+        "consensus_timestamp": "1234567890000000003",
+        "payerAccountId": "0.0.8",
+        "recipientAccountId": "0.0.9",
+        "amount": 30
+      }
+    ]
+  },
   "url": "/api/v1/transactions",
   "responseStatus": 200,
   "responseJson": {
     "transactions": [
       {
-        "consensus_timestamp": "0.000001050",
-        "valid_start_timestamp": "0.000001049",
-        "charged_tx_fee": 0,
+        "consensus_timestamp": "1234567890.000000003",
+        "valid_start_timestamp": "1234567890.000000002",
+        "charged_tx_fee": 7,
         "memo_base64": null,
         "result": "SUCCESS",
         "name": "CRYPTOTRANSFER",
-        "node": "0.15.2",
-        "transaction_id": "0.15.10-0-000001049",
+        "node": "0.0.3",
+        "transaction_id": "0.0.8-1234567890-000000002",
         "valid_duration_seconds": "11",
         "max_fee": "33",
         "transfers": [
           {
-            "account": "0.15.2",
+            "account": "0.0.3",
             "amount": 1
           },
           {
-            "account": "0.15.9",
-            "amount": 10
+            "account": "0.0.8",
+            "amount": -30
           },
           {
-            "account": "0.15.10",
-            "amount": -11
+            "account": "0.0.8",
+            "amount": -5
+          },
+          {
+            "account": "0.0.8",
+            "amount": -2
+          },
+          {
+            "account": "0.0.9",
+            "amount": 30
+          },
+          {
+            "account": "0.0.98",
+            "amount": 2
+          },
+          {
+            "account": "0.0.98",
+            "amount": 4
           }
         ]
       },
       {
-        "consensus_timestamp": "0.000001051",
-        "valid_start_timestamp": "0.000001050",
-        "charged_tx_fee": 0,
+        "consensus_timestamp": "1234567890.000000002",
+        "valid_start_timestamp": "1234567890.000000001",
+        "charged_tx_fee": 7,
         "memo_base64": null,
         "result": "SUCCESS",
         "name": "CRYPTOTRANSFER",
-        "node": "0.15.2",
-        "transaction_id": "0.15.10-0-000001050",
+        "node": "0.0.3",
+        "transaction_id": "0.0.10-1234567890-000000001",
         "valid_duration_seconds": "11",
         "max_fee": "33",
         "transfers": [
           {
-            "account": "0.15.2",
+            "account": "0.0.3",
             "amount": 1
           },
           {
-            "account": "0.15.9",
+            "account": "0.0.9",
             "amount": 20
           },
           {
-            "account": "0.15.10",
-            "amount": -21
+            "account": "0.0.10",
+            "amount": -20
+          },
+          {
+            "account": "0.0.10",
+            "amount": -5
+          },
+          {
+            "account": "0.0.10",
+            "amount": -2
+          },
+          {
+            "account": "0.0.98",
+            "amount": 2
+          },
+          {
+            "account": "0.0.98",
+            "amount": 4
           }
         ]
       },
       {
-        "consensus_timestamp": "0.000001052",
-        "charged_tx_fee": 0,
+        "consensus_timestamp": "1234567890.000000001",
+        "valid_start_timestamp": "1234567890.000000000",
+        "charged_tx_fee": 7,
         "memo_base64": null,
         "result": "SUCCESS",
         "name": "CRYPTOTRANSFER",
-        "node": "0.15.2",
-        "transaction_id": "0.15.8-0-000001051",
+        "node": "0.0.3",
+        "transaction_id": "0.0.10-1234567890-000000000",
+        "valid_duration_seconds": "11",
         "max_fee": "33",
         "transfers": [
           {
-            "account": "0.15.2",
+            "account": "0.0.3",
             "amount": 1
           },
           {
-            "account": "0.15.8",
-            "amount": -31
+            "account": "0.0.9",
+            "amount": 10
           },
           {
-            "account": "0.15.9",
-            "amount": 30
+            "account": "0.0.10",
+            "amount": -10
+          },
+          {
+            "account": "0.0.10",
+            "amount": -5
+          },
+          {
+            "account": "0.0.10",
+            "amount": -2
+          },
+          {
+            "account": "0.0.98",
+            "amount": 2
+          },
+          {
+            "account": "0.0.98",
+            "amount": 4
           }
         ],
         "valid_duration_seconds": "11",
-        "valid_start_timestamp": "0.000001051"
+        "valid_start_timestamp": "1234567890.000000000"
       }
     ],
     "links": {
