+++ hedera-mirror-rest/package.json
@@ -60,7 +60,8 @@
     "testPathIgnorePatterns": [
       "__tests__/testutils.js",
       "__tests__/mockpool.js",
-      "__tests__/integrationDbOps.js"
+      "__tests__/integrationDbOps.js",
+      "__tests__/integrationDomainOps.js"
     ]
   },
   "jest-junit": {
